refactor(workbench): extract sidebar context menus into WorkbenchFloatingMenus
Move project/session/composer context-menu item builders out of AppWorkbench so menu rows can change without opening the 18k-line host. Open/archive verbs stay on the host.
